Finished API, fixed modal messages
I finished patching the api, I also completed the get from api actions.
I was also able to display the results, but they are very ugly.
I will fix this in a future iteration.
I also cleared the display message in the modal after auth.
next steps are to clean up some of the css and some modal work.